# docs/api/openapi.yml
openapi: 3.0.0
info:
  title: Docketeer Extension API
  description: API for managing Docketeer resources.
  version: 1.0.0
paths:
...
  /api/docker/container/{id}/inspect:
    get:
      tags:
        - container
      summary: Inspect a container by ID
      parameters:
      - name: id
        in: path
        required: true
        description: Container ID
      responses:
        '200':
          description: Container details
          content:
            application/json:
              schema:
                $ref: 
  /api/docker/container/start:
    post:
      tags:
        - container
      summary: Start a container
      requestBody:
        description: Container start parameters
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                containerId:
                  type: string
                  description: ID of the container to start
      responses:
        '200':
          description: Container started
  /api/docker/container/stop:
    post:
      tags:
        - container
      summary: Stop a container
      requestBody:
        description: Container stop parameters
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                containerId:
                  type: string
                  description: ID of the container to stop
      responses:
        '200':
          description: Container stopped
  /api/docker/container/{id}:
    delete:
      tags:
        - container
      summary: Delete a container by ID
      parameters:
      - name: id
        in: path
        required: true
        description: Container ID
      responses:
        '204':
          description: Container deleted

  # Image endpoints
...
  /api/docker/network:
    get:
      tags:
        - network
      summary: List all networks
      responses:
        '200':
          description: List of networks
    delete:
      tags:
        - network
      summary: Delete all networks
      responses:
        '204':
          description: Networks deleted
    post:
      tags:
        - network
      summary: Create a network
      requestBody:
        description: Network creation parameters
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                networkName:
                  type: string
                  description: Name of the network to create
      responses:
        '201':
          description: Network created
...
  /api/docker/network/{id}/container/{containerId}:
    delete:
      tags:
        - network
      summary: Remove container from network by IDs
      parameters:
      - name: id
        in: path
        required: true
        description: Network ID
      - name: containerId
        in: path
        required: true
        description: Container ID
      responses:
        '204':
          description: Container removed from network

  # Volume endpoints
...